feat: refactored components names
